YG-131 🐞 게시글 상세 페이지 주소 및 type 이름 변경
