+++ apis/mockPosts.ts
@@ -5,7 +5,7 @@ export const samplePosts: Partial<Post>[] = [
         postId: 1,
         title: "룰루랄라 프랑스 여행 중",
         continent: "유럽",
-        region: "프랑스",
+        country: "프랑스",
         author: "Amy",
         createdAt: "2024-06-12",
         modifiedAt: "2024-06-12",
@@ -19,7 +19,7 @@ export const samplePosts: Partial<Post>[] = [
         postId: 2,
         title: "나는 국내가 좋아",
         continent: "아시아",
-        region: "한국",
+        country: "한국",
         author: "Gang",
         createdAt: "2024-06-12",
         modifiedAt: "2024-06-12",
@@ -33,7 +33,7 @@ export const samplePosts: Partial<Post>[] = [
         postId: 3,
         title: "일본 갔다 왔어요!",
         continent: "아시아",
-        region: "일본",
+        country: "일본",
         author: "Kimi",
         createdAt: "2024-06-12",
         modifiedAt: "2024-06-12",
@@ -47,7 +47,7 @@ export const samplePosts: Partial<Post>[] = [
         postId: 4,
         title: "미국 여행 중",
         continent: "북아메리카",
-        region: "미국",
+        country: "미국",
         author: "Wendy",
         createdAt: "2024-06-12",
         modifiedAt: "2024-06-12",
@@ -61,7 +61,7 @@ export const samplePosts: Partial<Post>[] = [
         postId: 5,
         title: "겨울에 너무 추워",
         continent: "유럽",
-        region: "노르웨이",
+        country: "노르웨이",
         author: "Jayden",
         createdAt: "2024-06-12",
         modifiedAt: "2024-06-12",
@@ -75,7 +75,7 @@ export const samplePosts: Partial<Post>[] = [
         postId: 6,
         title: "우유니 사막 최고",
         continent: "남아메리카",
-        region: "볼리비아",
+        country: "볼리비아",
         author: "Louis",
         createdAt: "2024-06-12",
         modifiedAt: "2024-06-12",
@@ -89,7 +89,7 @@ export const samplePosts: Partial<Post>[] = [
         postId: 7,
         title: "휴양은 바다지",
         continent: "아시아",
-        region: "코타키나발루",
+        country: "코타키나발루",
         author: "DK",
         createdAt: "2024-06-12",
         modifiedAt: "2024-06-12",
@@ -103,7 +103,7 @@ export const samplePosts: Partial<Post>[] = [
         postId: 8,
         title: "특이한 곳 여행이 취미",
         continent: "아프리카",
-        region: "이집트",
+        country: "이집트",
         author: "Jun",
         createdAt: "2024-06-12",
         modifiedAt: "2024-06-12",
@@ -117,7 +117,7 @@ export const samplePosts: Partial<Post>[] = [
         postId: 9,
         title: "열기구가 보고 싶어",
         continent: "아시아",
-        region: "터키",
+        country: "터키",
         author: "Amy",
         createdAt: "2024-06-12",
         modifiedAt: "2024-06-12",
@@ -131,7 +131,7 @@ export const samplePosts: Partial<Post>[] = [
         postId: 10,
         title: "동물 보러 옴",
         continent: "아프리카",
-        region: "가나",
+        country: "가나",
         author: "Gang",
         createdAt: "2024-06-12",
         modifiedAt: "2024-06-12",
