+++ apis/postApi.ts
@@ -55,7 +55,7 @@ export const postPost = async (newPost: CreatePost): Promise<CreatePost> => {
 }
 
 /**
- * @function putFreePost free-form으로 작성한 게시글의 수정
+ * @function putPost 작성한 게시글의 수정
  * @param {number} postId 수정할 게시글의 ID
  * @returns {Promise<UpdatePost>} 수정된 post의 내용을 객체로 반환
  */
@@ -71,57 +71,12 @@ export const putPost = async (postId: number, editedPost: UpdatePost): Promise<U
         address: editedPost.address || "",
         memos: editedPost.memos || [],
         continent: editedPost.continent || "",
-        region: editedPost.region || "",
+        country: editedPost.country || "",
         tripStartDate: editedPost.tripStartDate || "",
         tripEndDate: editedPost.tripEndDate || "",
         themeList: editedPost.themeList || [],
     }
 }
-
-/**
- * @function putFreePost free-form으로 작성한 게시글의 수정
- * @param {number} postId 수정할 게시글의 ID
- * @returns {Promise<UpdatePost>} 수정된 post의 내용을 객체로 반환
- */
-// export const putFreePost = async (postId: number, editedPost: UpdatePost): Promise<UpdatePost> => {
-//     const response = await fetchFormAPI(POST_API_URL, `posts/${postId}`, {
-//         method: "PUT",
-//         body: JSON.stringify(editedPost),
-//     })
-//     if (!response.ok) throw new Error("free-form 게시글 수정에 실패했어요...🥹")
-//     return {
-//         title: editedPost.title || "",
-//         content: editedPost.content || "",
-//         address: editedPost.address,
-//         continent: editedPost.continent || "",
-//         region: editedPost.region || "",
-//         tripStartDate: editedPost.tripStartDate || "",
-//         tripEndDate: editedPost.tripEndDate || "",
-//         themeList: editedPost.themeList || [],
-//     }
-// }
-
-/**
- * @function putMemoPost memo-form으로 작성한 게시글의 수정
- * @param {number} postId 수정할 게시글의 ID
- * @returns {Promise<Partial<UpdatePost>>} 수정된 post의 내용을 객체로 반환
- */
-// export const putMemoPost = async (shortPostId: number, editedPost: UpdatePost): Promise<UpdatePost> => {
-//     const response = await fetchFormAPI(POST_API_URL, `posts/short-posts/${shortPostId}`, {
-//         method: "PUT",
-//         body: JSON.stringify(editedPost),
-//     })
-//     if (!response.ok) throw new Error("memo-form 게시글 수정에 실패했어요...🥹")
-//     return {
-//         title: editedPost.title || "",
-//         memos: editedPost.memos || [],
-//         continent: editedPost.continent || "",
-//         region: editedPost.region || "",
-//         tripStartDate: editedPost.tripStartDate || "",
-//         tripEndDate: editedPost.tripEndDate || "",
-//         themeList: editedPost.themeList || [],
-//     }
-// }
 
 /**
  * @function deletePost 특정 id의 게시글을 삭제하는 함수
