+++ app/(afterLogin)/createPost/freeForm/page.tsx
@@ -1,9 +1,7 @@
-"use client"
-
 import CommonPost from "../_components/commonPost"
 
 const Page = () => {
-    return <CommonPost mode={"create"} isFreeForm={true} memos={[]} />
+    return <CommonPost isFreeForm={true} memos={[]} />
 }
 
 export default Page
