+++ app/(afterLogin)/createPost/memoForm/page.tsx
@@ -15,19 +15,22 @@ const Page = () => {
         const updatedMemos = memos.filter((_, i) => i !== index)
         setMemos(updatedMemos)
     }
-
     const handleEditorInputChange = (index: number, value: string) => {
         const updatedMemos = memos.map((shortPost, i) => (i === index ? { ...shortPost, content: value } : shortPost))
+        setMemos(updatedMemos)
+    }
+    const handleAddressInputChange = (index: number, field: string, value: string) => {
+        const updatedMemos = memos.map((shortPost, i) => (i === index ? { ...shortPost, [field]: value } : shortPost))
         setMemos(updatedMemos)
     }
 
     return (
         <CommonPost
-            mode={"create"}
             memos={memos}
             handleAddMemoClick={handleAddMemoClick}
             handleDeleteQuillEditor={handleDeleteQuillEditor}
             handleEditorInputChange={handleEditorInputChange}
+            handleAddressInputChange={handleAddressInputChange}
             isFreeForm={false}
         />
     )
