+++ utils/reset.utils.ts
@@ -6,7 +6,7 @@ export const getDefaultPost = (): CreatePost => ({
     address: "",
     memos: [],
     continent: "",
-    region: "",
+    country: "",
     tripEndDate: "",
     tripStartDate: "",
     themeList: [],
@@ -18,7 +18,7 @@ export const getUpdateDefaultPost = (postDetail: Post): UpdatePost => ({
     address: postDetail.address || "",
     memos: postDetail.memos || [],
     continent: postDetail.continent || "",
-    region: postDetail?.region || "",
+    country: postDetail?.country || "",
     tripEndDate: postDetail?.tripEndDate || "",
     tripStartDate: postDetail?.tripStartDate || "",
     themeList: postDetail?.themeList || [],
