+++ utils/search.utils.ts
@@ -21,19 +21,19 @@ export const filterPosts = (posts: Post[], searchTerm: string): Post[] => {
         const author = post.author ? post.author.toLowerCase() : ""
         const content = post.content ? post.content.toLowerCase() : ""
         const continent = post.continent ? post.continent.toLowerCase() : ""
-        const region = post.region ? post.region.toLowerCase() : ""
+        const country = post.country ? post.country.toLowerCase() : ""
 
         return (
             chosungIncludes(title, term) ||
             chosungIncludes(author, term) ||
             chosungIncludes(content, term) ||
             chosungIncludes(continent, term) ||
-            chosungIncludes(region, term) ||
+            chosungIncludes(country, term) ||
             title.includes(term) ||
             author.includes(term) ||
             content.includes(term) ||
             continent.includes(term) ||
-            region.includes(term)
+            country.includes(term)
         )
     })
 }
